requirements.txt
python-dotenv
playwright
flask
